% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{my descriptive package for data check and analysis}
\usage{
desc(prices)
}
\arguments{
\item{prices}{the stock prices of companies}
}
\value{
the descriptive statistics and a file excel
}
\description{
my descriptive package for data check and analysis
}
