% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{export}
\alias{export}
\title{This function is used to export to xls}
\usage{
export(y)
}
\arguments{
\item{y}{the describtive statistics output from \code{\link{desc}} function}
}
\value{
Excel file output with name descstatistics in the working directory
}
\description{
This function is used to export to xls
}
