% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdesc.R
\name{fdesc}
\alias{fdesc}
\title{My Descriptive Package for Data Check and Analysis}
\usage{
fdesc(prices)
}
\arguments{
\item{prices}{is the stock prices of datframe of firms}
}
\value{
the descriptive statistics and a file excel
}
\description{
My Descriptive Package for Data Check and Analysis
}
