% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{Rdesc}
\alias{Rdesc}
\title{Create table of Descriptive statistics}
\usage{
Rdesc(x)
}
\arguments{
\item{x}{is an xts data data with multiple variables}
}
\value{
the descriptive statistics for the log return
}
\description{
Create table of Descriptive statistics
}
