% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{My Descriptive Package for Data Check and Analysis}
\usage{
desc(x)
}
\arguments{
\item{x}{is an xts data data with multiple variables}
}
\value{
the descriptive statistics as a prepared table
}
\description{
My Descriptive Package for Data Check and Analysis
}
