% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{My Descriptive Package for Data Check and Analysis}
\usage{
desc(x)
}
\arguments{
\item{x}{is the stock prices of xts data}
}
\value{
the descriptive statistics and a file excel
}
\description{
My Descriptive Package for Data Check and Analysis
}
