% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{fdesc}
\alias{fdesc}
\title{my descriptive package for data check and analysis}
\usage{
fdesc(prices)
}
\arguments{
\item{prices}{the stock prices of companies}
}
\value{
the descriptive statistics and a file excel
}
\description{
my descriptive package for data check and analysis
}
